Fake errors generation and testing
Summary:
Adding bindings for the generation of fake errors on the apiserver
side, and testing them on the client side. This allows intentionally
simulating rare error paths in the DXHTTPRequest function, which is
foundational to dx. There are equivalent HTTP handlers in the other
language clients.

Lint fixes

Test Plan: Updated the test_dxclient.py test.

Reviewers: psung

Differential Revision: https://dnanexus.phacility.com/D718